% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rdf.R
\name{download_rdf}
\alias{download_rdf}
\title{Download new British National Bibliography records in RDF format}
\usage{
download_rdf(file_location, rdf_dates = c(""))
}
\arguments{
\item{file_location}{character string of path to directory where RDF file(s) will be stored}

\item{rdf_dates}{user-defined dates in the format d/m/y for which RDF files they want to download. If no argument is supplied, the most recent 25 records are downloaded}
}
\value{
RDF files in the user-defined directory
}
\description{
Download new British National Bibliography records in RDF format
}
\examples{
file_dates <- c("28/7/2021", "4/8/2021")
download_rdf("raw data/rdf", file_dates)
}
