% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_bnb.R
\name{query_bnb}
\alias{query_bnb}
\title{Query the RDF data and return as a tibble}
\usage{
query_bnb(rdf_data)
}
\arguments{
\item{rdf_data}{character vector containing path(s) to rdf file(s)}
}
\value{
tibble containing query result
}
\description{
Query the RDF data and return as a tibble
}
