% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rdf.R
\name{download_rdf}
\alias{download_rdf}
\title{Download new British National Bibliography records in RDF format}
\usage{
download_rdf(file_location, files_to_download = c(TRUE))
}
\arguments{
\item{file_location}{character string of path to directory where the file(s) will be stored}

\item{files_to_download}{user-defined dates in the format d/m/y for which RDF files they want to download. If no argument is supplied, the most recent 25 records are downloaded}
}
\value{
RDF files in the user-defined directory, formatted as .rdf.gzip
}
\description{
Download new British National Bibliography records in RDF format
}
\examples{
file_dates <- c("28/7/2021", "4/8/2021")
download_rdf("raw data/rdf", file_dates)
}
